Separated Hits controllers components
