+++ Sources/Widget/UIKit/Selectable/SelectableFilterButtonController.swift
@@ -4,7 +4,7 @@
 //
 //  Created by Vladislav Fitc on 03/05/2019.
 //  Copyright © 2019 Algolia. All rights reserved.
-//
+//b
 
 import Foundation
 import InstantSearchCore
